% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.BIS_BAS}
\alias{abcddata_add.BIS_BAS}
\title{Add Data for the BIS/BAS Scale}
\usage{
abcddata_add.BIS_BAS(
  dtf_ABCD_long_form,
  chr_files = "mh_y_bisbas.csv",
  chr_paths = "core/mental-health",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data with responses
to the BIS/BAS items.}

\item{chr_paths}{A character vector with
the associated folder path for the
BIS/BAS responses file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add individual items and summed scores
for subscales of the BIS/BAS scale. The ABCD® study
used a modified 20-item version of the scale developed by
Pagliaccio et al. (2016). There were 4 subscales
(BIS with 7 items, BAS: Reward Responsiveness with 5 items,
BAS: Drive with 4 items, and BAS: Fun Seeking with 4 items)
with a response format of 0 (Not true) to 3 (Very true).
The modified scale incorporated the BAS: Fun Seeking
subscale from the original measure by Carver and White (1994).
See Barch et al. (2018) for further details on the use of
the BIS/BAs scale in the ABCD® study.
}
\references{
Barch, D. M., Albaugh, M. D., Avenevoli, S., Chang, L., Clark,
D. B., Glantz, M. D., Hudziak, J. J., Jernigan, T. L.,
Tapert, S. F., Yurgelun-Todd, D., Alia-Klein, N., Potter,
A. S., Paulus, M. P., Prouty, D., Zucker, R. A., & Sher,
K. J. (2018). Demographic, physical and mental health
assessments in the Adolescent Brain and Cognitive Development
study: Rationale and description. Developmental Cognitive
Neuroscience, 32, 55–66. https://doi.org/10.1016/j.dcn.2017.10.010

Carver, C. S., and White, T. I. (1994). Behavioral inhibition,
behavioral activation and affective responses to impending
reward and punishment: The BIS/BAS scales. Journal of
Personality and Social Psychology, 67 (2), 319-333.

Pagliaccio, D., Luking, K. R., Anokhin, A. P., Gotlib, I. H.,
Hayden, E. P., Olino, T. M., Peng, C.-Z., Hajcak, G., &
Barch, D. M. (2016). Revising the BIS/BAS Scale to study
development: Measurement invariance and normative effects
of age and sex from childhood through adulthood.
Psychological Assessment, 28 (4), 429–442.
https://doi.org/10.1037/pas0000186
}
