% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.CBCL}
\alias{abcddata_add.CBCL}
\title{Add Data for the Child Behavior Checklist}
\usage{
abcddata_add.CBCL(
  dtf_ABCD_long_form,
  chr_files = "mh_p_cbcl.csv",
  chr_paths = "core/mental-health",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data with responses
to the CBCL items.}

\item{chr_paths}{A character vector with
the associated folder path for the
CBCL responses file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add individual items and summed scores
for subscales of the Child Behavior Checklist (CBCL;
Achenbach, 2009) for ages 6-18. The ABCD® study administered
112 items (note item 56 is broken into 8 sub-questions)
with a response format of 0 (Not true) to 2 (Very
true/often true). The study excluded the final
113th item which which involved free text entry. There
were 8 syndrome subscales (Anxious/Depressed with 13 items,
Withdrawn/Depressed with 8 items, Somatic Complaints with
11 items, Social Problems with 11 items, Thought Problems
with 15 items, Attention Problems with 10 items,
Rule-Breaking Behavior with 17 items, and
Aggressive Behavior with 18 items). Five of the
subscales group into two higher order factors: Internalizing
(Anxious/Depressed, Withdrawn/Depressed, and Somatic
Complaints) and Externalizing (Rule-Breaking Behavior,
Aggressive Behavior). A total score can also be computed
using all 112 items including the 16 items not included
in any subscale. The questions can also be
split into 6 subscales oriented for the Diagnostic and
Statistical Manual of Mental Disorders (Affective Problems,
Anxiety Problems, Somatic Problems, Attention-Deficit-Hyperactive
Disorder, Oppositional Defiant Problems, Conduct Problems).
See Barch et al. (2018) for further details on the use of
the CBCL scale in the  ABCD® study.
}
\references{
Achenbach, T. M. (2009). The Achenbach System of Empirically Based
Assessment (ASEBA): Development, findings, theory and applications.
University of Vermont, Research Center for Children, Youth, and
Families.

Barch, D. M., Albaugh, M. D., Avenevoli, S., Chang, L., Clark,
D. B., Glantz, M. D., Hudziak, J. J., Jernigan, T. L.,
Tapert, S. F., Yurgelun-Todd, D., Alia-Klein, N., Potter,
A. S., Paulus, M. P., Prouty, D., Zucker, R. A., & Sher,
K. J. (2018). Demographic, physical and mental health
assessments in the Adolescent Brain and Cognitive Development
study: Rationale and description. Developmental Cognitive
Neuroscience, 32, 55–66. https://doi.org/10.1016/j.dcn.2017.10.010
}
