% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.family_history}
\alias{abcddata_add.family_history}
\title{Add Data for Family History}
\usage{
abcddata_add.family_history(
  dtf_ABCD_long_form,
  chr_files = "mh_p_fhx.csv",
  chr_paths = "core/mental-health",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data on family history.}

\item{chr_paths}{A character vector with
the associated folder path for the
family history file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add data on assorted family history
measures (e.g., substance use problems, mental health
issues, etc.) for the ABCD® study.
}
