% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.NIH_toolbox}
\alias{abcddata_add.NIH_toolbox}
\title{Add Data for the NIH Toolbox}
\usage{
abcddata_add.NIH_toolbox(
  dtf_ABCD_long_form,
  chr_files = c("nc_y_nihtb.csv"),
  chr_paths = "core/neurocognition",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data from the NIH Toolbox.}

\item{chr_paths}{A character vector with
the associated folder path for the
NIH Toolbox file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add results from the National
Institute of Health (NIH) Toolbox for the ABCD®
study. Luciana et al. (2018) describe the specific
tasks from the toolbox used in the study.
}
\details{
The NIH toolbox provides multiple cognitive tasks for
participants to complete.
\enumerate{
\item The flanker task: A trial in the flanker task has
a child indicate whether a central stimulus (an
arrow) points left or right. The central stimulus can either
be between stimuli pointing in the same direction
(congruent trials) or the opposite direction (incongruent
trials). Following 4 practice trials, children completed
20 trials with 12 congruent trials and 8 incongruent trials.
Slotkin et al. (2012) provide details on scoring, though
note the details are for the full 40 trial version of the task that
includes an initial set of 20 trials using fish instead
of arrows intended for children younger than 8 years of age.
}
}
\references{
Luciana, M., Bjork, J.M., Nagel, B.J., Barch, D.M., Gonzalez,
R., Nixon, S.J., Banich, M.T. (2018). Adolescent neurocognitive
development and impacts of substance use: Overview of the
Adolescent Brain Cognitive Development (ABCD) baseline
neurocognition battery. Dev. Cogn. Neuroscience, 32, 67–79.
https://doi.org/10.1016/j.dcn.2018.02.006

Slotkin, J., Nowinski, C., Hays, R., Beaumont, J., Griffith, J.,
Magasi, S., Salsman, J., & Gershon, R. (2012). NIH Toolbox
scoring and interpretation guide. National Institutes of Health.
}
