% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.sample_characteristics}
\alias{abcddata_add.sample_characteristics}
\title{Add Data for Sample Characteristics}
\usage{
abcddata_add.sample_characteristics(
  dtf_ABCD_long_form,
  chr_files = "abcd_p_demo.csv",
  chr_paths = "core/abcd-general",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data on demographics.}

\item{chr_paths}{A character vector with
the associated folder path for the
demographics file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add data on sample characteristics
for the ABCD® study. The function extracts
variables regarding a child's sex at birth,
current gender identity, ethnicity, and
race (per the 2015 NIH standards). The function
also extracts variables regarding the
combined income, employment status, marital
status, and highest education obtained for the
child's parents/guardians.
}
