% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.substance_use_perceived_harm}
\alias{abcddata_add.substance_use_perceived_harm}
\title{Add Data for Perceived Harm of Substance Use}
\usage{
abcddata_add.substance_use_perceived_harm(
  dtf_ABCD_long_form,
  chr_files = "su_y_percharm.csv",
  chr_paths = "core/substance-use",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data on family history.}

\item{chr_paths}{A character vector with
the associated folder path for the
family history file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add data from 16 items on the perceived
harm of substance use, adapted from the Monitoring the
Future Study (e.g., Johnston et al., 2017) for the ABCD®
study.
}
\references{
Johnston, L. D., O'Malley, P. M., Miech, R. A., Bachman,
J. G., Schulenberg, J. E. (2017). Monitoring the Future
national survey results on drug use, 1975–2016: Overview,
key findings on adolescent drug Use. Institute for Social
Research, The University of Michigan.
}
