% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.UPPS}
\alias{abcddata_add.UPPS}
\title{Add Data for the UPPS-P Scale}
\usage{
abcddata_add.UPPS(
  dtf_ABCD_long_form,
  chr_files = "mh_y_upps.csv",
  chr_paths = "core/mental-health",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the data with responses
to the UPPS-P items.}

\item{chr_paths}{A character vector with
the associated folder path for the
UPPS-P responses file.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add individual items and summed scores
for subscales of the UPPS-P Impulsive Behavior
Scale for Children.
}
\details{
The ABCD® study used a modified
20 item version of the scale developed by
Zapolski et al. (2010). There were 5 subscales
(Negative urgency, perseverance, premeditation,
sensation seeking, and positive urgency)
consisting of 4 items each on with a response format
of 1 (Not at all like me) to 4 (Very much like me)
coded so higher scores always indicate greater
impulsivity.

See Barch et al. (2018) for further
details on the use of the UPPS-P scale in the
ABCD® study.
}
\references{
Barch, D. M., Albaugh, M. D., Avenevoli, S., Chang, L., Clark,
D. B., Glantz, M. D., Hudziak, J. J., Jernigan, T. L.,
Tapert, S. F., Yurgelun-Todd, D., Alia-Klein, N., Potter,
A. S., Paulus, M. P., Prouty, D., Zucker, R. A., & Sher,
K. J. (2018). Demographic, physical and mental health
assessments in the Adolescent Brain and Cognitive Development
study: Rationale and description. Developmental Cognitive
Neuroscience, 32, 55–66. https://doi.org/10.1016/j.dcn.2017.10.010

Zapolski, T. C. B., Stairs, A. M., Settles, R. F., Combs,
J. L., & Smith, G. T. (2010). The measurement of dispositions
to rash action in children. Assessment, 17 (1), 116-125.
https://doi.org/10.1177/1073191109351372
}
