% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_add_entry}
\alias{abcddata_codebook_add_entry}
\title{Add a Codebook Entry to a Data Frame Column}
\usage{
abcddata_codebook_add_entry(
  dtf_base,
  chr_column,
  chr_description = "",
  lst_values_and_labels = NULL,
  lst_collected_over = NULL,
  chr_source_files = "",
  chr_source_variables = "",
  chr_notes = "",
  chr_category = "",
  chr_data_type = "",
  chr_reporter = ""
)
}
\arguments{
\item{dtf_base}{A data frame.}

\item{chr_column}{A character string, the column name
of the variable for which a codebook entry will be added.}

\item{chr_description}{A character string, a human-readable
description of the variable and what it is for.}

\item{lst_values_and_labels}{A list of two vectors,
the values and their associated labels if applicable.}

\item{lst_collected_over}{A list of two vectors giving the
levels and associated variables over which the given
variable was collected.}

\item{chr_source_files}{A character vector, the source
files from which the variable was taken.}

\item{chr_source_variables}{A character vector, the
original names of the source variables from which
the variable was derived.}

\item{chr_notes}{A character vector, additional
notes on the variable.}

\item{chr_category}{A character string, the category type
for the variable (see \link{abcddata_codebook_column_types}).}

\item{chr_data_type}{A character string, the data type
for the variable (see \link{abcddata_codebook_column_types}).}

\item{chr_reporter}{A character string, the responder type
for the variable (see \link{abcddata_codebook_column_types}).}
}
\value{
A data frame with the associated column updated with
a new attribute.
}
\description{
Function to update a column of a data frame to
have a codebook entry.
}
\examples{
dtf_base <- data.frame(
  IDS.CHR.GD.Participant = LETTERS[1:3],
  QST.INT.CS.Example = 4:6
)

dtf_base <- abcddata_codebook_add_entry(
  dtf_base, 'IDS.CHR.GD.Participant',
  chr_description = 'Participant identifier'
)
abcddata_codebook_display( dtf_base$IDS.CHR.GD.Participant )

dtf_base <- abcddata_codebook_add_entry(
  dtf_base, 'QST.INT.CS.Example',
  chr_description = 'Example questionnaire responses'
)
abcddata_codebook_display( dtf_base$QST.INT.CS.Example )

}
