% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_column_types}
\alias{abcddata_codebook_column_types}
\title{Column Types for ABCD Codebook Entries}
\usage{
abcddata_codebook_column_types(chr_column_name = "", chr_abbr = NULL)
}
\arguments{
\item{chr_column_name}{A character string,
the column name to process. If the
empty character set \code{''}, lists
the abbreviations and associated labels
for each abbreviation type.}

\item{chr_abbr}{A character string, the
abbreviation type to consider, either
\code{'Category'}, \code{'Data type'},
or \code{'Responder'}.}
}
\value{
Either a character string, the label
associated with the abbreviation of interest,
or displays a message to the console listing
abbreviations and labels.
}
\description{
Function to convert abbreviations used in a
column name into the associated label to
include in a codebook entry for ABCD data.
}
\examples{

abcddata_codebook_column_types()

chr_column <- 'IDS.CHR.GD.Subject'
abcddata_codebook_column_types( chr_column, 'Category' )
abcddata_codebook_column_types( chr_column, 'Data type' )
abcddata_codebook_column_types( chr_column, 'Reporter' )

}
