% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_content}
\alias{abcddata_codebook_content}
\title{List of Content for Codebook Entries}
\usage{
abcddata_codebook_content(vec_content, vec_additional_content)
}
\arguments{
\item{vec_content}{A vector, must match in length
to \code{'vec_additional_content'}.}

\item{vec_additional_content}{A vector, must match
in length to \code{'vec_content'}.}
}
\value{
A list of two vectors of matching length
labeled 'content' and 'additional_content'
respectively.
}
\description{
Function to create an appropriately labeled
list with the elements 'content' and
'additional_content' for use with codebook
entries.
}
\examples{
abcddata_codebook_content( 1:3, c( 'A', 'B', 'C' ) )

}
