% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_display}
\alias{abcddata_codebook_display}
\title{Display ABCD Codebook Entry}
\usage{
abcddata_codebook_display(vec_values, int_text_wrap = 60)
}
\arguments{
\item{vec_values}{A vector with an \code{'abcd_codebook'}
element in the attributes.}
}
\value{
As a side effect, displays the codebook entry
in the console window.
}
\description{
Function to display an ABCD codebook entry
contained in a vector's attributes.
}
