% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_extract}
\alias{abcddata_codebook_extract}
\title{Extract ABCD Codebook Entry}
\usage{
abcddata_codebook_extract(vec_values)
}
\arguments{
\item{vec_values}{A vector with an \code{'abcd_codebook'}
element in the attributes.}
}
\value{
A data frame.
}
\description{
Function to extract the data frame with the
ABCD codebook entry contained in a vector's
attributes.
}
