% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-codebook.R
\name{abcddata_codebook_new_entry}
\alias{abcddata_codebook_new_entry}
\title{Create a New Codebook Entry}
\usage{
abcddata_codebook_new_entry(
  chr_variable_name = "",
  chr_category = "",
  chr_data_type = "",
  chr_reporter = "",
  chr_description = "",
  lst_values_and_labels = NULL,
  lst_collected_over = NULL,
  chr_source_files = "",
  chr_source_variables = "",
  chr_notes = ""
)
}
\arguments{
\item{chr_variable_name}{A character string, the column
name for the variable.}

\item{chr_category}{A character string, the category type
for the variable (see \link{abcddata_codebook_column_types}).}

\item{chr_data_type}{A character string, the data type
for the variable (see \link{abcddata_codebook_column_types}).}

\item{chr_reporter}{A character string, the reporter
for the variable (see \link{abcddata_codebook_column_types}).}

\item{chr_description}{A character string, a human-readable
description of the variable and what it is for.}

\item{lst_values_and_labels}{A list of two vectors,
the values and their associated labels if applicable.}

\item{lst_collected_over}{A list of two vectors giving the
levels and associated variables over which the given
variable was collected.}

\item{chr_source_files}{A character vector, the source
files from which the variable was taken.}

\item{chr_source_variables}{A character vector, the
original names of the source variables from which
the variable was derived.}

\item{chr_notes}{A character vector, additional
notes on the variable.}
}
\value{
A data frame.
}
\description{
Function to create a new codebook entry
for a ABCD data set variable.
}
\examples{
abcddata_codebook_new_entry()

}
