% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-utilities.R
\name{abcddata_propagate}
\alias{abcddata_propagate}
\title{Propagate Values Over Multiple Columns}
\usage{
abcddata_propagate(
  dtf_data,
  chr_over,
  chr_columns,
  lgc_subset = NULL,
  lgc_progress = FALSE
)
}
\arguments{
\item{dtf_data}{A data frame.}

\item{chr_over}{A character string, the
column to use as the index variable.}

\item{chr_columns}{A character vector, the
columns for whom values should be propagated.}

\item{lgc_subset}{An optional logical vector
equal in length to the number of rows of
\code{dtf_data}, the subset of rows to consider.}

\item{lgc_progress}{Logical; if \code{TRUE} displays the
progress of the function for debugging purposes.}
}
\value{
A data frame, equivalent to \code{dtf_data} but
with propagated values for the columns listed in
\code{chr_columns}.
}
\description{
Function to propagate a value across
multiple rows based on an index variable.
}
\examples{
dtf_example <- data.frame(
  ID = rep( 1:3, each = 3 ),
  X = c( 1, NA, NA, 2, NA, NA, 3, NA, NA ),
  Y = c( NA, NA, 'A', NA, NA, 'B', NA, NA, 'C' )
)

dtf_example <- abcddata_propagate(
  dtf_example,
  chr_over = 'ID',
  chr_columns = c( 'X', 'Y' )
)

}
