% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-utilities.R
\name{abcddata_replace}
\alias{abcddata_replace}
\title{Replace Values in a Vector}
\usage{
abcddata_replace(
  vec_values,
  vec_to_replace,
  vec_replace_with,
  obj_default = NA,
  lgc_exact = TRUE
)
}
\arguments{
\item{vec_values}{A vector of values.}

\item{vec_to_replace}{A vector, the unique elements in
\code{vec_values} to replace.}

\item{vec_replace_with}{A vector (equal in length to
\code{vec_to_replace}), the new elements to replace those
specified in \code{vec_to_replace}.}

\item{obj_default}{The default value to use for elements
in \code{vec_values} that aren't updated.}

\item{lgc_exact}{Logical; if \code{TRUE} elements in
\code{vec_to_replace} must exactly match those in \code{vec_values},
otherwise uses \link[base:grep]{base::grepl} for partial pattern matching.}
}
\value{
A vector equal in length to \code{vec_values} with
updated values.
}
\description{
Function to replace values in a vector. Robust against
\code{NA} values.
}
\examples{
chr_example <- rep( LETTERS[1:3], each = 2 )
abcddata_replace( chr_example, c( 'A', 'B' ), c( 1, 2 ) )

}
