% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-utilities.R
\name{\%with\%}
\alias{\%with\%}
\title{Operator to Match Patterns in Strings}
\usage{
chr_vector \%with\% chr_pattern
}
\arguments{
\item{chr_vector}{A character vector.}

\item{chr_pattern}{A character string,
the pattern to match against.}
}
\value{
A logical vector equal in length
to \code{chr_vector}, set to \code{TRUE} when
the given string contains \code{chr_pattern}.
}
\description{
Operator to find all elements in a
character vector that contain a specified
pattern. Makes a call to \link[base:grep]{base::grepl},
}
\examples{
c( 'cat', 'bat', 'dog' ) \%with\% 'at'

}
