% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-package.R
\docType{package}
\name{abcddata-package}
\alias{abcddata-package}
\title{Data Preparation for the ABCD® Study}
\description{
The \pkg{abcddata} package provides assorted
functions to help with processing data obtained
from the Adolescent Brain Cognitive Development (ABCD)
Study (\url{https://abcdstudy.org/}), held in the NIMH Data
Archive (NDA). See \url{https://rettopnivek.github.io/abcddata/}
for more details regarding the R package.
}
