% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcddata-data_creation.R
\name{abcddata_add.substance_use}
\alias{abcddata_add.substance_use}
\title{Add Data on Substance Use}
\usage{
abcddata_add.substance_use(
  dtf_ABCD_long_form,
  chr_files = "su_y_sui.csv",
  chr_paths = "core/substance-use",
  lgc_progress = TRUE
)
}
\arguments{
\item{dtf_ABCD_long_form}{A data frame, output from
the \link{abcddata_initialize_long_form}
function.}

\item{chr_files}{A character vector with
the file name for the substance use data.}

\item{chr_paths}{A character vector with
the associated folder path for the
substance use data.}

\item{lgc_progress}{A logical value; if
\code{TRUE} displays the progress of the
data processing.}
}
\value{
A data frame.
}
\description{
Function to add data on whether children engaged
in any substance use.
}
\details{
Two type of substance use
are defined as per Sullivan et al. (2022):
experimentation (low-level use, such as alcohol
sipping, nicotine/cannabis puffing or trying)
versus initiation (one or more standard drinks
of alcohol, more than a puff or taste of
nicotine/cannabis, and any kind of use of
other types of substances).

Substances were categorized into four types:
alcohol, nicotine or tobacco, cannabis, and
other substances (which consisted of: synthetic MJ,
sniffing products to get
high, prescription med abuse, cocaine, bath salts,
meth, ecstasy, ketamine, GBH, heroin, hallucinogens,
mushrooms, salvia, or steroids). See Lisdahl et al.
(2018) for more details on how the ABCD® study collected
data on substance use.
}
\references{
Lisdahl, K. M., Sher, K. J., Conway, K. P., Gonzalez, R.,
Feldstein Ewing, S. W., Nixon, S. J., Tapert, S., Bartsch,
H., Goldstein, R. Z., & Heitzeg, M. (2018). Adolescent
brain cognitive development (ABCD) study: Overview of
substance use assessment methods. Developmental
Cognitive Neuroscience, 32, 80–96.
https://doi.org/10.1016/j.dcn.2018.02.007

Sullivan, R. M., Wade, N. E., Wallace, A. L., Tapert,
S. F., Pelham, W. E., Brown, S. A., Cloak, C. C.,
Ewing, S. W. F., Madden, P. A. F., Martz, M. E.,
Ross, J. M., Kaiver, C. M., Wirtz, H. G., Heitzeg,
M. M., & Lisdahl, K. M. (2022). Substance use patterns
in 9 to 13-year-olds: Longitudinal findings from the
Adolescent Brain Cognitive Development (ABCD) study.
Drug and Alcohol Dependence Reports, 5, 1–12.
https://doi.org/10.1016/j.dadr.2022.100120
}
